mapi: remove shebang line
The scripts should be invoked as an argument to the appropriate python
implementation. That's what the scons and autoconf builds to upstream.

My guess is that's also the case for the legacy MSVC project files used
here, since there is no /usr/bin/env on Windows ;-)

Patch supplied by Emil Velikov <emil.l.velikov@gmail.com>